Abstracted explicit solvation functionality from endo_ace_ex into explicit_solvation.py
